added popup on prenotazione ombrellone

+++ Casotto-Front-End/src/app/bar/riga-bar/riga-bar.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { RigaCatalogoBar } from '../riga-catalogo-bar';
+import { RigaCatalogoBar } from 'src/app/model/riga-catalogo-bar'; 
 import { RigheBarService } from '../righe-bar.service';
 
 import { ActivatedRoute, Router } from '@angular/router';
@@ -12,29 +12,30 @@ import { Location } from '@angular/common';
 })
 export class RigaBarComponent implements OnInit {
 
-  public righe? : RigaCatalogoBar[];
+  public righe?: RigaCatalogoBar[];
 
-  title : string = "Catalogo Bar"
+  title: string = "Catalogo Bar"
 
   constructor(private righeBarService: RigheBarService,
-              private route: ActivatedRoute,  
-              private location: Location,     
-              private router: Router) { }
-
-  public getRigheBar() : void {
-    this.righeBarService.getRighe().subscribe(r => this.righe = r)
-  }
-
-  public rigaBarAggiunta(rigaAggiunta : RigaCatalogoBar) {
-    this.righe?.push(rigaAggiunta)
-  }
-
-  eliminaRiga(rigaDaEliminare : RigaCatalogoBar) {
-    this.righeBarService.removeRiga(rigaDaEliminare.id).subscribe(() => this.getRigheBar())
-  }
+    private route: ActivatedRoute,
+    private location: Location,
+    private router: Router) { }
 
   ngOnInit(): void {
     this.getRigheBar()
+  }
+
+  public getRigheBar(): void {
+    this.righeBarService.getRighe().subscribe(r => this.righe = r)
+  }
+
+  public rigaBarAggiunta(rigaAggiunta: RigaCatalogoBar) {
+    this.righe?.push(rigaAggiunta)
+  }
+
+  eliminaRiga(rigaDaEliminare: RigaCatalogoBar) {
+    if (this.righeBarService.askConfirm("eliminare", "eliminato"))
+      this.righeBarService.removeRiga(rigaDaEliminare.id).subscribe(() => this.getRigheBar())
   }
 
 }
